Create SealDownloader.
